refactor(surrogates): remove complicated surrogate system
This commit removes a bunch of dead code which in the original version of this codebase was used to
support having various different kinds of surrogate data. Now that there is really only one we've
stabilized on (the "Randomized Spike Matrix") and this method has been incorporated upstream into
braingeneerspy, there is no reason to support anything other than selecting between real and RSM
data in hmmsupport.