feat(exploratory): answers to Tal's questions
StateTraversal now also includes box plots of the average entropy of the
state distribution, weighted mainly towards the burst peak (but results
are similar if all time points are weighted evenly).

There is also a plot of the number of distinct trajectories, but since
there is no statistically significant difference, this is less obviously
useful, just something Tal was curious about.