feat(exploratory): HMM consistency of scaffold units
Add the script HMMBackbone.py, which uses hypothesis testing to check
how Poisson the behavior of each unit in the recording is. The hope is
that scaffold/packet/backbone units will be significantly lower-variance
than the Poisson hypothesis, but the result isn't yet as clear-cut as I
hoped it would be.

Also reformat PCAComparison with black.