feat(backbone): more nuanced consistency scores
Two different modifications are included here, both of which make the distributions of consistency scores significantly more separable and are also quite easy to justify:
1. Only consider the consistency of units during states corresponding to bursts.
2. Stop counting NaN states as "indistinguishable from Poisson".

Theoretically the second should be a much more important change, but it didn't have as much of an effect as I was expecting. If anything, it actually broadened out the distributions quite a bit. This may actually be an important step of controlling for firing rate... I haven't decided yet if it's more correct to do it this way or the old way.