fix(backbone): broken handling of empty sequences
Previously np.full_like(data) was being used to generate NaN p-values for cases where data included
no samples. The problem is that data was integer-valued, so NaN was invalid! It was silently
converted to a very negative integer, which then became a very negative float, which in turn caused
zero-data cases to be treated as significant! This seriously skewed my results, and the
distributions look much saner now...