fix(nrp): miscellaneous cv worker issues
Modify `Model.cross_validate()` to expose the `verbose` parameter, and
have the CV worker use it, so it's clear what's going on.

The `justfile` needed a new recipe to actually run the queueing
implemented by `do_cv.py`, which in turn got a few cosmetic tweaks.

Also make the worker loop be the one to print which jobs are received.