fix(supplements): compute CV scores like Maboudi
The paper that originally motivated this HMM analysis didn't implement
cross-validation the way I had it in the original draft. I was training
models on both real and surrogate data and comparing their log
likelihood on their respective validation sets, but it's not really
correct to compare log likelihood when both the model and the dataset
are different.

This commit includes an implementation of a new cross validation
function based on the one included in SSM that trains the model on a CV
fold and evaluates it on a validation set, but also reports the
performance of the same subset of the surrogate data.